# specs/001-rebuild-music-parser/contracts/api-contract.yaml
openapi: 3.1.0
info:
  title: Music Parser API Contract
  version: 1.1.0
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/downloads/{jobId}:
    get:
      operationId: downloadJobArtifact
      parameters:
        - $ref: '#/components/parameters/JobId'
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MP3 file stream
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
...
components:
  parameters:
    JobId:
      in: path
      name: jobId
      required: true
      schema:
        type: string
        format: uuid
...